requirements.txt
requests
Pillow
rembg
pytesseract
python-dotenv
tenacity
openai
